requirements.txt
langgraph
langchain-groq
langchain-community
duckduckgo-search
chromadb
streamlit
python-dotenv